Time: 0 ms (100%), Space: 8.5 MB (92.36%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    double solve(double x, long n){
+    double solve(double x, long long n){
         if(n==0) return 1;
         double half= solve(x,n/2);
         if(n%2==0){//even
